Fix issue where an empty password would be interpreted as an anonymous bind
If anonymous binds were allowed on the LDAP server and an empty
password was used, valid? would return true.  This is a very
serious bug, and the only two workarounds are to disable
anonymous binds in ldap or disallow empty passwords completely.
As this library has no control over the server, I chose to
disallow empty passwords completely.

Bump version to 1.0.1.